fix: Accept EOF in non-interactive mode anywhere to end board creation
The issue was when reading from stdin, we didn't know if it's interactive and want double newline to finish the board creation, or if it's coming from a file redirect and we accept EOF anywhere.

I added a function which should reliably detect if `STDIN_FILENO` is interactive or not, with the limited functions we are allowed to use.

With this I was also able to get rid of the global `g_stdin` variable.